+++ packages/dummy/src/component/dummy/__tests__/Dummy.stories.ts
@@ -1,8 +1,8 @@
-import { Dummy } from '../Dummy';
+import { Dummy } from '@musica-sacra/dummy';
 import { StoryObj } from '@storybook/react';
 
 const meta = {
-    title: 'dummy/Dummy',
+    title: '@musica-sacra/Dummy',
     component: Dummy,
     args: {
         dummyText: 'Dummy text',