+++ packages/dummy/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @musica-sacra/dummy
 
+## 0.0.3
+
+### Patch Changes
+
+- b21c866: Fix story title
+
+## 0.0.2
+
+### Patch Changes
+
+- 7cac3b8: Add example change to dummy text.
+  - **What** Add example change to dummy text
+  - **Why** Want to release new patch version
+  - **How** Nothing.
+
 ## 0.0.1
 
 ### Patch Changes